+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Added
 
+- `PathfindingSearch#done()` - check if search is complete
+- `PathfindingSearch#resultBlocking()` - block until result is available
+- `PathfindingSearch#result()` - get result if available
+
 ### Changed
 
 ### Fixed